chore: Move timings structs to dim-independent crate
